requirements.txt
glfw==2.9.0
networkx==3.4.2
numpy==2.2.5
pillow==11.2.1
pyglm==2.8.2
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
